1. Added nominal_time parameter.
This is needed for PK analysis and allows the option of including a nominal relative time from first dose variable (e.g. NFRLT).

The value of nominal time will be updated with each row according to the values of EXDOSFRQ so "QD" will result in increments of 24 hours and "BID" will result in increments of 12 hours for example.

2. Fixed an issue where times for "DAY" values if present were not updated even if the interval was less than a whole day, i.e. "BID", "TID" and "QID".  For example, "Q12H" is being handled differently from "BID".